% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kb_click.R
\name{kb_click}
\alias{kb_click}
\alias{kb_focus}
\title{Keybindings for Shiny Widgets}
\usage{
kb_click(inputId, key = NULL, modifier = NULL)

kb_focus(inputId, key = NULL, modifier = NULL)
}
\arguments{
\item{inputId}{inputID of the widget you want to click on}

\item{key}{the keyvalue e.g. "a". See Details}

\item{modifier}{Modifier Key. Allowed values are "ctrl", "alt" (option or ⌥ on OS X) or "shift". Only one modifier is allowed.}
}
\description{

}
\details{
Other common keys you might want to use besides the usal alphanumerical keys:
\itemize{
\item 'Enter' for '\code{'Enter'}, \code{'↵'} or \code{'Return'}
}

https://jsfiddle.net/4j54jqt2/4
}
\section{Functions}{
\itemize{
\item \code{kb_focus}: Focus on a widget
}}

\examples{
\dontrun{shinykeybindings::kb_examples()}
}
